% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_group_interaction.R
\name{boxplot_group_interaction}
\alias{boxplot_group_interaction}
\title{Create a boxplot of the joint mean score for a specific ligand receptor
interaction between two groups of samples}
\usage{
boxplot_group_interaction(
  ligand,
  receptor,
  cell_type1,
  cell_type2,
  sample_replicates,
  sample_group_1,
  sample_group_2
)
}
\arguments{
\item{ligand}{Ligand of interest}

\item{receptor}{Receptor of interest}

\item{cell_type1}{Cell type expressing the ligand of interest}

\item{cell_type2}{Cell type expressiong the receptor of interest}

\item{sample_replicates}{Name of the meta.data column in a Seurat object that
has the samples of the individual replicate samples}

\item{sample_group_1}{The group of samples from which interactions will be
identified for statistical comparison}

\item{sample_group_2}{Comparitor group from which interactions from
sample_group_1 will be measured against}
}
\value{
A ggplot boxplot of the joint mean for a ligand/receptor interaction
in each sample group.
}
\description{
Create a boxplot of the joint mean score for a specific ligand receptor
interaction between two groups of samples
}
