% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltalk.R
\name{celltalk}
\alias{celltalk}
\title{Assess ligand and receptor interactions across groups of cells}
\usage{
celltalk(
  input_object,
  metadata_grouping = cell_types,
  ligand_receptor_pairs = ramilowski_pairs,
  number_cells_required = 100,
  min_expression = 1000,
  max_expression = 20000,
  scramble_times = 10
)
}
\arguments{
\item{input_object}{Seurat object to create ligand receptor matrices from}

\item{metadata_grouping}{Grouping variable in the Seurat object metadata used
to define the groups of cells. Default is "cell_types".}

\item{ligand_receptor_pairs}{Data.frame of ligands, receptors and
interactions in the format of ramilowski_pairs provided by this package.
Defaults is "ramilowski_pairs".}

\item{number_cells_required}{Number of cells per group required to perform
analysis of ligand/receptor interactions. Defaults to 100.}

\item{min_expression}{Minimum expression in counts to consider a ligand or
receptor for interactions analysis. A sensible default is set to 1000, but is
dataset dependent. This is meant to filter out lowly expressed ligands and
receptors.}

\item{max_expression}{Maxmium expression in counts to consider a ligand or
receptor for interactions analysis. A sensible default is set to 20000, but is
dataset dependent. This is meant to filter out ubiquitously expressed ligands
and receptors.}

\item{scramble_times}{Number of times to scamble ligand/receptor interactions to
create a background distribution for statistical comparison.}
}
\value{
Comprehensive tibble of cognate ligand and receptor interactions and
statistical significance of these interactions across cell types.
}
\description{
Assess ligand and receptor interactions across groups of cells
}
