% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_plot.R
\name{circos_plot}
\alias{circos_plot}
\title{Creates a circos plot from the list of ligands and receptors}
\usage{
circos_plot(interactions1, interactions2 = NULL, clusters,
  ligand.col = "#C8A757", receptor.col = "#70BC71",
  interactions.col = "lightgray",
  unique.vs.control.interactions.col = "darkorange")
}
\arguments{
\item{interactions1}{List of ligands and receptors and their associated cell types (i.e. the output from one sample group from the putative_interactions function). If interactions2 is NULL, then this will only plot interactions found in the list of ligands and receptors from the sample group specified.}

\item{interactions2}{List of ligands and receptors and their associated cell types (i.e. the output from a second sample group from the putative_interactions function). Default is NULL, but if this value is non-null, will plot both sets of interactions present in "interactions1" and "interactions2". This group is considered the "control" group if present, and unique ligand/receptor interactions from "interactions1" will be plotted.}

\item{ligand.col}{Color of the bars defining the ligands}

\item{receptor.col}{Color of the bars definig the receptors}

\item{interactions.col}{Color of the links connecting ligands and receptors. Default is light gray}

\item{unique.vs.control.interactions.col}{Optional color of the links connecting the ligand and receptor interactions that are unique between the sample group and the control}
}
\value{
Generates a circos plot connecting ligands and receptors across cell types for a given sample group
}
\description{
Creates a circos plot from the list of ligands and receptors
}
