% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_group_interactions.R
\name{compare_group_interactions}
\alias{compare_group_interactions}
\title{Statistically compare interactions between groups of samples}
\usage{
compare_group_interactions(
  interactions_stats,
  sample_replicates,
  sample_group_1,
  sample_group_2
)
}
\arguments{
\item{sample_replicates}{Name of the meta.data column in a Seurat object that
has the samples of the individual replicate samples}

\item{sample_group_1}{The group of samples from which interactions will be
identified for statistical comparison}

\item{sample_group_2}{Comparitor group from which interactions from
sample_group_1 will be measured against}

\item{interaction_stats}{Tibble from celltalk function, usually filtered to
the top significant ligand and receptor interactions of interest}
}
\value{
A data.frame contain ligands, receptors, cell types and the
statistical significant of the joint means in sample_group_1 versus
the joint means in sample_group_2
}
\description{
Statistically compare interactions between groups of samples
}
