% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lig_rec_tib.R
\name{create_lig_rec_tib}
\alias{create_lig_rec_tib}
\title{Create a tibble of consistently expressed ligands and receptors}
\usage{
create_lig_rec_tib(exp.tib, clusters, groups, replicates, cells.reqd,
  freq.pos.reqd, ligands.and.receptors)
}
\arguments{
\item{exp.tib}{The nested tibble that was the output from \code{reshape_matrices}.}

\item{clusters}{A named vector containing the cluster identities of each cell as a factor, where the names match the column names from the count matrix.}

\item{groups}{A named vector containing the group identities of each cell as a factor, where the names match the columna names from the count matrix.}

\item{replicates}{A named vector containing the replicate identities of each sample as a factor, where the names match the column names from the count matrix. Assumes there are biological replicates of each sample.}

\item{cells.reqd}{An integer value representing the number of cells required to express a given ligand or receptor within a cluster from an individual replicate.}

\item{freq.pos.reqd}{A fraction from 0 to 1 representing the fraction of individual replicate samples that must express a ligand or receptor above the \code{cells.reqd} threshold to be counted as a ligand or receptor in a cluster from a given group.}

\item{ligands.and.receptors}{A matrix with 3 columns, the first of which corresponds to ligands, the second of which corresponds to receptors and the last of which corresponds to ligand_receptor interactions. Column names should be "ligand", "receptor", and "pairs" respectively.}
}
\value{
Returns a nested tibble containing the ligands and receptors consistently expressed within clusters from each group.
}
\description{
Create a tibble of consistently expressed ligands and receptors
}
