% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_interactions.R
\name{diff_interactions}
\alias{diff_interactions}
\title{Evaluate differential interactions}
\usage{
diff_interactions(putative.interactions.tib, group1, group2, cluster,
  search.ligand = TRUE, search.receptor = FALSE, ligands.and.receptors)
}
\arguments{
\item{putative.interactions.tib}{A tibble output from \code{putative_interactions}.}

\item{group1}{Sample group to evaluate for unqiue ligands or receptors participating in interactions.}

\item{group2}{Sample group as a comparitor to group1}

\item{cluster}{A cluster to evaluate unique ligand/receptor interactions between groups.}

\item{search.ligand}{Logical dictating whether to search for the presence of a unique ligand between groups. Must be TRUE if \code{search.receptor} is FALSE.}

\item{ligands.and.receptors}{A matrix with 3 columns, the first of which corresponds to ligands, the second of which corresponds to receptors and the last of which corresponds to ligand_receptor interactions. Column names should be "ligand", "receptor", and "pairs" respectively.}

\item{searh.receptor}{Logical dictating whether to search for the presence of a unique ligand between groups. Must be TRUE if \code{search.ligand} is FALSE.}
}
\value{
Returns a tibble containing the interactions where a cluster has a unique ligand or receptor interaction.
}
\description{
Evaluates whether a given cluster participates in putative ligand/receptor interactions between groups. Can search for a cluster with a ligand or receptor participating in differential interactions.
}
