% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_network.R
\name{diff_network}
\alias{diff_network}
\title{Differential network analysis for putatitve ligand receptor interactions between groups}
\usage{
diff_network(putative.interactions.tib, group1, group2, interaction)
}
\arguments{
\item{putative.interactions.tib}{A tibble output from \code{putative_interactions}.}

\item{group1}{Sample group to evaluate for unqiue ligands or receptors participating in interactions.}

\item{group2}{Sample group as a comparitor to group1}

\item{interaction}{Name of a ligand_receptor interaction to evaluate}
}
\value{
Returns a tibble containing the differential graphs for group1 versus group2 and group2 versus group1, and the differential edges from group1 versus group2 and group2 versus group1. The differential edges are derived from the igraph package.
}
\description{
Takes the output tibble from \code{putative_intearctions} and compares the networks of putative ligand receptor interactions between groups.
}
