% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filtered_lig_rec}
\alias{filtered_lig_rec}
\title{Expressed ligands and receptors.}
\format{
A sparse matrix with 192 genes (rows) and 23734 cells (columns)
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE139324/}
}
\usage{
filtered_lig_rec
}
\description{
A dataset containing the filtered expressed of ligands and receptors from
a series of peripheral blood from 5 healthy donors and tonsil from 5
healthy donors undergoing tonsilectomy.
}
\keyword{datasets}
