% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligand_receptor_means.R
\name{ligand_receptor_means}
\alias{ligand_receptor_means}
\title{Assess joint means between ligands and receptors across cell types}
\usage{
ligand_receptor_means(
  ligand_mean_dataframe,
  receptor_mean_dataframe,
  ligand_receptor_pairs,
  number_cell_types
)
}
\arguments{
\item{ligand_mean_dataframe}{Mean expression values for ligands across cells
from a given condition}

\item{receptor_mean_dataframe}{Mean expression values for receptors across
cells from a given condition}

\item{ligand_receptor_pairs}{Data.frame of ligands, receptors and
interactions in the format of ramilowski_pairs provided by this package}

\item{number_cell_types}{The number of unique cell types in a given sample
or condition. This will be used to assess ligand and receptor interactions
across cell types.}
}
\value{
Generates a tibble containing joint means for ligand receptor
interaction pairs across cell types.
}
\description{
Assess joint means between ligands and receptors across cell types
}
