% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligand_receptor_scramble.R
\name{ligand_receptor_scramble}
\alias{ligand_receptor_scramble}
\title{Assess joint means between scrambled interaction pairs for
ligands and receptors across cell types}
\usage{
ligand_receptor_scramble(
  ligand_receptor_means,
  ligand_mean_dataframe,
  receptor_mean_dataframe,
  ligand_receptor_pairs,
  number_cell_types,
  scramble_times
)
}
\arguments{
\item{ligand_receptor_means}{Tibble generated by ligand_receptor_means
function. Used to generate scrambled values to test statistical signifiance
of the strength of interactions.}

\item{ligand_mean_dataframe}{Mean expression values for ligands across cells
from a given condition}

\item{receptor_mean_dataframe}{Mean expression values for receptors across
cells from a given condition}

\item{ligand_receptor_pairs}{Data.frame of ligands, receptors and
interactions in the format of ramilowski_pairs provided by this package}

\item{number_cell_types}{The number of unique cell types in a given sample
or condition. This will be used to assess ligand and receptor interactions
across cell types.}

\item{scramble_times}{Number of times to random pair ligands and receptors to
generate a background distribution.}
}
\value{
Generates a tibble containing joint resulting from scrambling ligand
and receptor pairs.
}
\description{
Assess joint means between scrambled interaction pairs for
ligands and receptors across cell types
}
