% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligand_receptor_stats.R
\name{ligand_receptor_stats}
\alias{ligand_receptor_stats}
\title{Assess statistical significance of ligand/receptor interactions within an
individual group of samples}
\usage{
ligand_receptor_stats(ligand_receptor_means, ligand_receptor_scramble)
}
\arguments{
\item{ligand_receptor_means}{Tibble generated by ligand_receptor_means
function. Used to generate create a joint expression weighting for ligand
receptor pairs.}

\item{ligand_receptor_scramble}{Tibble generated by ligand_receptor_scramble
function. Used to generate scrambled values to test statistical signifiance
of the strength of interactions.}
}
\value{
Generates a tibble containing joint resulting from scrambling ligand
and receptor pairs.
}
\description{
Assess statistical significance of ligand/receptor interactions within an
individual group of samples
}
