% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overall_metadata}
\alias{overall_metadata}
\title{Metadata from 5 health blood donors and 5 healthy tonsil donors.}
\format{
A data.frame with 23734 cells (rows) and 8 variables:
\describe{
  \item{orig.ident}{name assigned by default to the Seurat object}
  \item{nCount_RNA}{number of RNA molecules counted per cell (1038--74386)}
  \item{nFeature_RNA}{number of RNA molecules counted per cell (41--6582)}
  \item{sample_id}{individual donor samples, HD_PBMC 1 to 5 and HD_Tonsil 1 to 5}
  \item{sample_type}{tissue of origin, either PBMC or Tonsil}
  \item{RNA_snn_res.0.3}{Louvian clustering results, clusters 0 to 13}
  \item{seurat_clusters}{clusters assigned by default, same as RNA_snn_res.0.3}
  \item{cell_types}{cell types identified by canonical gene expression (B cells,
   CD14-CD16+ monocytes, CD14+CD16- monocytes, CD1C+ DCs, CD4 T cells,
   CD8 T cells, NK cells, pDCs, Plasmablasts, RBCs)}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE139324/}
}
\usage{
overall_metadata
}
\description{
Overall metadata from analysis of 5 healthy blood donors and 5 healthy
tonsil tissue donors.
}
\keyword{datasets}
