% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overall_umap}
\alias{overall_umap}
\title{UMAP of cells from 5 health blood donors and 5 healthy tonsil donors.}
\format{
A matrix with 23734 cells (rows) and 2 dimensions of UMAP embedding
(columns)
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE139324/}
}
\usage{
overall_umap
}
\description{
The UMAP_1 and UMAP_2 coordinates of cells resulting from dimensionality
reduction of all cells from 5 healthy peripheral blood donors and 5 tonsil
tissue donors.
}
\keyword{datasets}
