% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putative_interactions.R
\name{putative_interactions}
\alias{putative_interactions}
\title{Create a tibble of consistently expressed ligands and receptors}
\usage{
putative_interactions(ligand.receptor.tibble, clusters, groups,
  freq.group.in.cluster, ligands.and.receptors)
}
\arguments{
\item{ligand.receptor.tibble}{The nested tibble that was the output from \code{create_lig_rec_tib}.}

\item{clusters}{A named vector containing the cluster identities of each cell as a factor, where the names match the column names from the count matrix.}

\item{groups}{A named vector containing the group identities of each cell as a factor, where the names match the columna names from the count matrix.}

\item{freq.group.in.cluster}{A fraction from 0 to 1 describing the frequency of cells from a given group required to be in a cluster to consider that cluster for interactions.}

\item{ligands.and.receptors}{A matrix with 3 columns, the first of which corresponds to ligands, the second of which corresponds to receptors and the last of which corresponds to ligand_receptor interactions. Column names should be "ligand", "receptor", and "pair" respectively.}
}
\value{
Returns a nested tibble containing putative interactions between clusters for a given group, where each putative interaction is a member of a list and the clusters expressing cognate ligands and receptore are in the list.
}
\description{
Takes the tibble output from \code{create_lig_rec_tib} and compares ligand and receptor expression withn and between clusters within a given sample group. Uses interactions described in \code{ligands.and.receptors}.
}
