% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ramilowski_pairs}
\alias{ramilowski_pairs}
\title{List of known ligands, receptors and their iteractions.}
\format{
A data.frame with 2557 rows and 3 variables:
\describe{
  \item{ligand}{expressed ligands, total of 708 unique ligands}
  \item{receptor}{expressed receptors, total of 691 receptors}
  \item{pair}{interaction between and ligand and receptor, total of 2557
   unique interactions}
}
}
\source{
\url{https://www.nature.com/articles/ncomms8866}
}
\usage{
ramilowski_pairs
}
\description{
This is a curated list of ligands, receptors and their potential interactions.
The list is derived from a manuscript from Ramilowski et al. There are other
potential lists that could be used in place of or in addition to this list.
Notably, the CellPhoneDB (https://github.com/ventolab/CellphoneDB) package
has an extensively curated list including multiple receptor components.
}
\keyword{datasets}
