% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_matrices.R
\name{reshape_matrices}
\alias{reshape_matrices}
\title{Reshape matrices}
\usage{
reshape_matrices(count.matrix, clusters, groups, replicates,
  ligands.and.receptors)
}
\arguments{
\item{count.matrix}{A raw count matrix with rows as genes and columns as cells.}

\item{clusters}{A named vector containing the cluster identities of each cell as a factor, where the names match the column names from the count matrix.}

\item{groups}{A named vector containing the group identities of each cell as a factor, where the names match the columna names from the count matrix.}

\item{replicates}{A named vector containing the replicate identities of each sample as a factor, where the names match the column names from the count matrix. Assumes there are biological replicates of each sample.}

\item{ligands.and.receptors}{A matrix with 3 columns, the first of which corresponds to ligands, the second of which corresponds to receptors and the last of which corresponds to ligand_receptor interactions. Column names should be "ligand", "receptor", and "pair" respectively.}
}
\value{
Returns a nested tibble containing the individual expression matrices, with the first level as the groups and the second level as the individual replciates within each group.
}
\description{
Takes an overall count matrix and metadata describing the clusters, groups, replicates and ligands and receptors as input and generates individual gene expression matrices for each cluster within each replicate within each group. Returns a nested tibble with the individual expression matrices.
}
