% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_interactions.R
\name{unique_interactions}
\alias{unique_interactions}
\title{Unique interactions between two groups}
\usage{
unique_interactions(putative.interactions.tib, group1, group2,
  ligands.and.receptors)
}
\arguments{
\item{putative.interactions.tib}{A tibble output from \code{putative_interactions}.}

\item{group1}{Sample group to evaluate for unqiue ligands or receptors participating in interactions.}

\item{group2}{Sample group as a comparitor to group1}

\item{ligands.and.receptors}{A matrix with 3 columns, the first of which corresponds to ligands, the second of which corresponds to receptors and the last of which corresponds to ligand_receptor interactions. Column names should be "ligand", "receptor", and "pairs" respectively.}
}
\value{
Returns a tibble containing interactions unique to group1 versus group2, group2 versus group1 and interactions that are common between the groups.
}
\description{
Identifies interaction pairs that are exclusively present in one group versus another at the group level
}
