% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_group_interaction.R
\name{boxplot_group_interaction}
\alias{boxplot_group_interaction}
\title{Create a boxplot of the joint mean score for a specific ligand receptor
interaction between two groups of samples}
\usage{
boxplot_group_interaction(
  seurat_object,
  interaction_stats,
  sample_replicates,
  sample_groups,
  metadata_grouping,
  ligand,
  receptor,
  cell_type1,
  cell_type2
)
}
\arguments{
\item{seurat_object}{Seurat object containing expression levels across all
cells and sample types}

\item{interaction_stats}{Tibble from celltalk function, usually filtered to
the top significant ligand and receptor interactions of interest}

\item{sample_replicates}{Name of the meta.data column in a Seurat object that
has the samples of the individual replicate samples}

\item{sample_groups}{Name of the meta.data column in a Seurat object that
has the name of the sample group}

\item{metadata_grouping}{Name of the meta.data column in a Seurat object that
has the name of the groups of cells to evaluate (e.g. "cell_types" containing
previously identified cell types)}

\item{ligand}{Name of the ligand in the ligand/receptor pair of interest}

\item{receptor}{Name of the receptor in the ligand/receptor pair of interest}

\item{cell_type1}{Name of the cell type expressing the ligand in the ligand/
receptor pair of interest}

\item{cell_type2}{Name of the cell type expressing the receptor in the ligand/
receptor pair of interest}
}
\value{
A ggplot boxplot of the joint mean for a ligand/receptor interaction
in each sample group.
}
\description{
Create a boxplot of the joint mean score for a specific ligand receptor
interaction between two groups of samples
}
