% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_plot.R
\name{circos_plot}
\alias{circos_plot}
\title{Creates a circos plot from the list of ligands and receptors}
\usage{
circos_plot(ligand.receptor.frame, colors, lig.col, rec.col)
}
\arguments{
\item{ligand.receptor.frame}{List of ligands and receptors and their associated cell types (i.e. the output from one sample group from the putative_interactions function).}

\item{colors}{Color of the bars defining the ligands}

\item{lig.col}{Color of the bars definig the receptors}

\item{rec.col}{Color of the links connecting ligands and receptors. Default is light gray}
}
\value{
Generates a circos plot connecting ligands and receptors across cell types for a given sample group
}
\description{
Creates a circos plot from the list of ligands and receptors
}
