% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_group_interactions.R
\name{compare_group_interactions}
\alias{compare_group_interactions}
\title{Statistically compare interactions between groups of samples}
\usage{
compare_group_interactions(
  seurat_object,
  interaction_stats,
  sample_replicates,
  sample_groups,
  metadata_grouping
)
}
\arguments{
\item{seurat_object}{Seurat object containing expression data acorss all
samples and cells}

\item{interaction_stats}{Tibble from celltalk function, usually filtered to
the top significant ligand and receptor interactions of interest}

\item{sample_replicates}{Name of the meta.data column in a Seurat object that
has the samples of the individual replicate samples}

\item{sample_groups}{Name of the meta.data column in a Seurat object that
has the name of the sample group}

\item{metadata_grouping}{Name of the meta.data column in a Seurat object that
has the name of the groups of cells to evaluate (e.g. "cell_types" containing
previously identified cell types)}
}
\value{
A list of linear models containing the coefficients and statistics for
assessing differences in joint ligand and receptor expression between sample
groups. Note that this requires replicate samples from each samples group.
}
\description{
Statistically compare interactions between groups of samples
}
